[refactor] 불필요한 import와 함수 리팩터링 & DB 타입 변경

+++ src/main/java/com/kernel360/kernelsquare/domain/member_answer_vote/entity/MemberAnswerVote.java
@@ -27,7 +27,7 @@ public class MemberAnswerVote extends BaseEntity {
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
-	@Column(nullable = false, name = "status", columnDefinition = "tinyint")
+	@Column(nullable = false, name = "status", columnDefinition = "smallint")
 	private Integer status;
 
 	@ManyToOne(fetch = FetchType.LAZY)
